requirements.txt
# Python dependencies for Notion export
python-dotenv==1.0.1
